requirements.txt
# Deep Learning Frameworks
torch==2.0.1
torchvision==0.15.2
transformers==4.30.0

# Data Processing
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.3.0
Pillow==9.5.0

# Visualization
matplotlib==3.7.1
seaborn==0.12.2
tensorboard==2.13.0

# Text Processing
nltk==3.8.1

# Image Processing
albumentations==1.3.1
opencv-python==4.8.0.74

# Utilities
tqdm==4.65.0
pyyaml==6.0

# CLIP (optional, for v0.4+)
ftfy==6.1.1
regex==2023.6.3
